refactor(Header): splits Header into multiple components
