+++ src/components/Header/Menu/Fullscreen/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MenuFullscreen.vue";
