+++ src/components/Header/Menu/Shrinked/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MenuShrinked.vue";
