+++ src/consts/contents.js
@@ -1,0 +1,18 @@
+export default [
+  {
+    id: "about",
+    translateKey: "menu.about"
+  },
+  {
+    id: "skills",
+    translateKey: "menu.skills"
+  },
+  {
+    id: "projects",
+    translateKey: "menu.projects"
+  },
+  {
+    id: "professional",
+    translateKey: "menu.professional"
+  }
+];
